% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{the amount of people to check the probability of sharing a birthday}
}
\value{
the probability that of x many people there is at least two people who share a birthday
}
\description{
Birthday
}
\examples{
birthday(20:24)
}
