% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire dataset}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from nearest fire station (in miles)}
  \item{DAMAGE}{Amount of fire damage (thousands of dollars)}
  ...
}
}
\source{
<Statistics for Engineering and the Sciences, 6th ed>
}
\usage{
fire
}
\description{
A dataset of information about fire damage levels and their distance from the nearest fire station.
}
\keyword{datasets}
