% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2
Bootstrap confidence interval for a sample statistic}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Integer, number of bootstrap resamples to draw. Default is 10000.}

\item{x}{Numeric vector containing the sample data.}

\item{fun}{Statistic to compute on each bootstrap sample.}

\item{alpha}{Significance level for the confidence interval.}

\item{cx}{Numeric scaling factor for text size in the plot. Default is 1.5.}

\item{...}{Additional graphical arguments passed to `hist()`.}
}
\value{
A list
}
\description{
myboot2
Bootstrap confidence interval for a sample statistic
}
\examples{
x <- rnorm(20, mean = 10, sd = 4)
res <- myboot2(iter = 2000, x = x, fun = "mean", alpha = 0.05)
res$ci
}
