% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{CLT demo}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample size.}

\item{iter}{Number of replicates.}

\item{a}{Uniform interval start.}

\item{b}{Uniform interval endpoint.}
}
\value{
A plot of a histogram with a CLT normal overlay and a numeric vector with the simulated means
}
\description{
CLT demo
}
\examples{
mycltu(n=10, iter=10000)
}
