% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstFunction.R
\name{myFirstFunction}
\alias{myFirstFunction}
\title{Title}
\usage{
myFirstFunction(x)
}
\arguments{
\item{x}{a quantitative vector}
}
\value{
A list with components 'x'
}
\description{
Title
}
\examples{
myFirstFunction(1:10)
}
