% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats available}

\item{gamma}{Acceptance rate of failure}

\item{p}{Probability that each passenger shows up}
}
\value{
A named list containing the discrete solution (nd), the continuous solution (nc), N, p, and gamma.
}
\description{
Calcuulates the optimal number of tickets to sell for a flight with N many seats. To allow airlines to sell the most tickets they can while managing overbooking risk.
Uses both a binomial method and a normal approximation.
}
\examples{
ntickets(400, 0.02, 0.95)
}
