% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssmssrss.R
\name{tssmssrss}
\alias{tssmssrss}
\title{Calculate Total, Model, and Residual Sum of Squares}
\usage{
tssmssrss(dataframe, quad.lm, col)
}
\arguments{
\item{dataframe}{A data frame containing the observed data.}

\item{quad.lm}{An object of class lm, typically a fitted quadratic regression model.}

\item{col}{A numeric vector of observed response values must align with quad.lm.}
}
\value{
A numeric vector of length 3 with elements: TSS, MSS, RSS
}
\description{
Calculate Total, Model, and Residual Sum of Squares
}
\examples{
df <- head(mtcars)  # take first 6 rows for simplicity

quad.lm <- lm(mpg ~ poly(hp, 2, raw = TRUE), data = df)

tssmssrss(df, quad.lm, df$mpg)
}
